[#185] refactor: common 분리
