+++ client/src/components/practice/long/TypingChar/index.tsx
